ConnMan: Do not set manager's offlinemode property if not needed
The switch toggled by an external event (like the extension or ConnMan
itself) triggers an unnecessary setting of the offlinemode property.
Avoid this by storing it, and comparing it with the switch state.
if both are same it was an externally triggered event in which case
we dont need to call the SetProperty method. if they are different then
its the UI triggering it, and we need to call the method on manager.