ConnMan: use separate text entries for each proxy server type
As ConnMan allows us to store any arbitrary string value as a proxy server
URL we need to take steps in the UI to encourage the user to enter
appropriate values. This change adds separate entries for each proxy
protocol, spin buttons for port numbers and uses placeholder text to
explain appropriate values.